moving trainNetwork.py to nn.trainUtils so it can be used outside of apps. Moving the distillery to its own project.
