+++ trunk/modules/python/nn/trainUtils.py
